refactor: rename to vue-input-facade
The code has significatly diverged from upstream, therefore I am renaming it and publishing it as
its own component.

BREAKING CHANGE: The component is now named input-facade, the directive and filter are now called
facade. The directive and filter can be changed via the `name` option passed to the plugin.