fix(core): cursor positioning when pasting, better detection of our own event
When pasting data into an empty field, if the value changed with masked characters the cursor would
sometimes end up mid value if the last digit pre-masked happen to also appear before the end of the
value.

Changed from Event to CustomEvent, so that we can set a flag to know that it is our own even rather
than relying on the isTrusted flag.  This allows the input to still be editable programatically:

  el.value = 'new value'
  el.dispatchEvent(new Event('input'))