fix(dynamic): changes the way we choose the mask from the dynamics array
Prior we were choosing the mask based on the size of the masked value and exited early if a shorter mask was encountered in the loop.  The logic is now updated to choose a mask based on the which one gives us the most information from the input. This also seems to fix the issue faced when using this library for dollar amounts.

fixes #1