fix(directive): get the current value from a component's model when available
When the directive is used on a custom component that has its own model, but that model is not been
bound back to the input's value property, the directive is not getting the proper value as part of
the update hook. Such is the case with vuetify's v-text-input.  This can be resolved by checking the
vnode to see if a model is defined, if so get the value for masking from the model rather than from
the input.value.

fixes #10, fixes #11, resolves vuetifyjs/vuetify#10890