fix(directive.js): fix handling events for multiple elements (#41)
Problem: If two input elements with the v-facade directive are both in the
same root element, then the event handlers for each will trigger for ALL
inputs on the elements. This can be problematic in cases where an input
event cannot or should not be handled by the masking function, ie
when one input is masked and the other is not.

Solution: Add a check to the event handler to confirm that the event
target is in fact the element that was initially target by the v-facade we care about.